Cut extra code
