+++ src/handlers/int-test-support/clients.ts
@@ -3,7 +3,7 @@ import { CloudWatchLogsClient } from "@aws-sdk/client-cloudwatch-logs";
 import { LambdaClient } from "@aws-sdk/client-lambda";
 import { S3Client } from "@aws-sdk/client-s3";
 import { SESClient } from "@aws-sdk/client-ses";
-import { KMS } from "aws-sdk";
+import AWS from "aws-sdk";
 
 export const region = "eu-west-2";
 
@@ -26,6 +26,6 @@ export const lambdaClient = new LambdaClient({ region });
 
 export const sesClient = new SESClient({ region: "eu-west-1" });
 
-export const kms = new KMS({
+export const kms = new AWS.KMS({
   region: "eu-west-2",
 });
