--- src/handlers/int-test-support/helpers/mock-data/invoice/data/street-types.json
@@ -1,1 +1,0 @@
-["Street", "Road", "Place", "Square", "Alley", "End"]
