+++ src/handlers/int-test-support/helpers/mock-data/invoice/data/street-types.ts
@@ -1,0 +1,1 @@
+export default ["Street", "Road", "Place", "Square", "Alley", "End"];
