+++ src/handlers/int-test-support/helpers/s3Helper.ts
@@ -10,7 +10,7 @@ import {
 import { runViaLambda } from "./envHelper";
 import { s3Client } from "../clients";
 import { sendLambdaCommand } from "./lambdaHelper";
-import { IntTestHelpers } from "../handler";
+import { IntTestHelpers } from "../types";
 import { callWithRetryAndTimeout } from "./call-wrappers";
 import { poll } from "./commonHelpers";
 import { decryptKms } from "./kmsHelper";
