+++ ui-tests/pageobjects/homepage.ts
@@ -1,4 +1,4 @@
-import Page from "./page.js";
+import Page from "./page";
 
 class HomePage extends Page {
   /**
